requirements.txt
google-genai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
Pillow>=10.0.0
boto3>=1.34.0
edge-tts>=6.1.9
